# auth/swagger/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "auth MX - OpenAPI 3.0",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://www.tavistoica.xyz"
    }
  ],
...
  "paths": {
...
    "/api/users/refresh-token": {
      "post": {
        "tags": ["auth MX"],
        "summary": "Generates a new Refresh Token to the user",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "refreshToken provided in the cookie header",
            "example": "refreshToken=eteoigtjo13j513;",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "Set-Cookie": {
                "description": "Contains the session cookie named `refreshToken`. Pass this cookie back in subsequent requests.\n",
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenSetCookie"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenSuccess"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authorization Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RefreshTokenSetCookie": {
        "type": "string",
        "example": "refreshToken=abcde12345; Path=/; HttpOnly"
      },
      "RefreshTokenSuccess": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["Customer", "Restaurant"]
          },
          "accessToken": {
            "type": "string"
          }
        },
        "required": ["role", "accessToken"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "description": "Optional text that describes the error in more detail",
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
    }
  }
}